deleted previous graphql folder and created resolvers and typedefs folders as replacements and put everything into respective folders, everything still works, some bugs in it but can be fixed later or work around with seperate queries
