+++ resolvers/index.js
@@ -1,0 +1,9 @@
+const bidResolver = require('./bidResolver');
+const companyResolver = require('./companyResolver');
+const studyResolver = require('./studyResolver');
+
+module.exports = [
+    bidResolver,
+    companyResolver,
+    studyResolver
+]
